Complete a rough first draft.
This `Nextflow` script will now take `cellranger count` outputs and
generate `AnnData` and `Seurat` objects. They are annotated with
various QC statistics as well as gene types. The pipeline also generates
plots of the QC statistics, while generating an `HTML` document that
summarizes the pipeline's outputs. This `HTML` summary is incomplete at
the moment and likely to change a lot as the outputs are organized.