Create a (nearly) fully functional annotation pipeline
This commit creates a nearly fully functional annotation pipeline.
After `cellranger count` creates a `filtered_feature_bc_matrix`, `soupx`
creates its own from the `cellranger` outputs. The `soupx` matrix should
have ambient RNA filtered out. Both matrices are then annotated in
parallel for various gene types and doublets, and plots are generated
automatically, as well as a web summary including plots. The web summary
is as of yet incomplete, as it does not contain a list of the genes
annotated, nor does it create the correct directory tree. The flowchart
is also incorrect at the moment.